s/janken_controller/cure_janken_manager/
